chore: 修复全部 62 个编译警告 → 0 warnings
改动 (7 files / 24 lines):

1. e1000.rs (8 warnings):
   - 移除未使用导入: inb, inw, outb, inw
   - _did, _bar_size_mask, _pkt_data, _netif → 变量前缀
   - /// → // (extern 块上 doc comment 无效)

2. netif.rs (2 warnings):
   - _data, _dev → 参数前缀

3. utils.rs (1 warning):
   - current_val: u8→u16 (避免 >255 永远为 false)

4. idt/mod.rs (1 warning):
   - addr! macro:  as *const () as usize as u64 (禁止 fn→整型直转)

5. syscall/mod.rs + handlers.rs (2+ warnings):
   - ata_disk_present: bool→i32 类型一致 (真实定义返回 i32)
   - 所有调用点: if func() → if func() != 0 / == 0

安全措施:
   - backup/ 目录 + .gitignore 忽略 (恢复用)

验证:
   - cargo check: 0 warnings ✓
   - QEMU: 正常启动 + IRQ0 持续触发 ✓
   - 合规: 804/804, 100% ✓

+++ src/kernel/net/utils.rs
@@ -352,12 +352,12 @@ pub fn format_ipv4(ip: [u8; 4], buf: &mut [u8]) -> usize {
 pub fn parse_ipv4(str: &[u8]) -> Option<[u8; 4]> {
     let mut result = [0u8; 4];
     let mut octet_idx = 0;
-    let mut current_val: u8 = 0;
+    let mut current_val: u16 = 0;
 
     for &ch in str {
         match ch {
             b'0'..=b'9' => {
-                current_val = current_val.wrapping_mul(10).wrapping_add(ch - b'0');
+                current_val = current_val.wrapping_mul(10).wrapping_add((ch - b'0') as u16);
                 if current_val > 255 {
                     return None;
                 }
@@ -366,7 +366,7 @@ pub fn parse_ipv4(str: &[u8]) -> Option<[u8; 4]> {
                 if octet_idx >= 3 {
                     return None;
                 }
-                result[octet_idx] = current_val;
+                result[octet_idx] = current_val as u8;
                 octet_idx += 1;
                 current_val = 0;
             }
@@ -378,7 +378,7 @@ pub fn parse_ipv4(str: &[u8]) -> Option<[u8; 4]> {
     if octet_idx != 3 {
         return None;
     }
-    result[octet_idx] = current_val;
+    result[octet_idx] = current_val as u8;
 
     Some(result)
 }
